+++ score-repo/src/main/java/org/oagi/score/repo/api/impl/jooq/businessterm/JooqBusinessTermReadRepository.java
@@ -65,12 +65,12 @@ public class JooqBusinessTermReadRepository
             businessTerm.setExternalReferenceId(record.get(BUSINESS_TERM.EXTERNAL_REF_ID));
             businessTerm.setExternalReferenceUri(record.get(BUSINESS_TERM.EXTERNAL_REF_URI));
             businessTerm.setCreatedBy(new ScoreUser(
-                    record.get(APP_USER.as("creator").APP_USER_ID.as("creator_user_id")).toBigInteger(),
+                    record.get(APP_USER.as("creator").APP_USER_ID.as("creator_user_id")),
                     record.get(APP_USER.as("creator").LOGIN_ID.as("creator_login_id")),
                     (byte) 1 == record.get(APP_USER.as("creator").IS_DEVELOPER.as("creator_is_developer")) ? DEVELOPER : END_USER
             ));
             businessTerm.setLastUpdatedBy(new ScoreUser(
-                    record.get(APP_USER.as("updater").APP_USER_ID.as("updater_user_id")).toBigInteger(),
+                    record.get(APP_USER.as("updater").APP_USER_ID.as("updater_user_id")),
                     record.get(APP_USER.as("updater").LOGIN_ID.as("updater_login_id")),
                     (byte) 1 == record.get(APP_USER.as("updater").IS_DEVELOPER.as("updater_is_developer")) ? DEVELOPER : END_USER
             ));
