+++ score-repo/src/main/java/org/oagi/score/repo/api/impl/jooq/module/JooqModuleSetReleaseReadRepository.java
@@ -69,7 +69,7 @@ public class JooqModuleSetReleaseReadRepository
             moduleSetRelease.setModuleSetReleaseName(record.get(MODULE_SET_RELEASE.NAME));
             moduleSetRelease.setModuleSetReleaseDescription(record.get(MODULE_SET_RELEASE.DESCRIPTION));
             moduleSetRelease.setModuleSetName(record.get(MODULE_SET.NAME));
-            moduleSetRelease.setReleaseId(record.get(MODULE_SET_RELEASE.RELEASE_ID).toBigInteger());
+            moduleSetRelease.setReleaseId(record.get(MODULE_SET_RELEASE.RELEASE_ID));
             moduleSetRelease.setReleaseNum(record.get(RELEASE.RELEASE_NUM));
             moduleSetRelease.setDefault(record.get(MODULE_SET_RELEASE.IS_DEFAULT) == 1);
 
@@ -78,11 +78,11 @@ public class JooqModuleSetReleaseReadRepository
             moduleSetRelease.setCreatedBy(
                     (isCreatorAdmin) ?
                             new ScoreUser(
-                                    record.get(APP_USER.as("creator").APP_USER_ID.as("creator_user_id")).toBigInteger(),
+                                    record.get(APP_USER.as("creator").APP_USER_ID.as("creator_user_id")),
                                     record.get(APP_USER.as("creator").LOGIN_ID.as("creator_login_id")),
                                     Arrays.asList(creatorRole, ADMINISTRATOR)) :
                             new ScoreUser(
-                                    record.get(APP_USER.as("creator").APP_USER_ID.as("creator_user_id")).toBigInteger(),
+                                    record.get(APP_USER.as("creator").APP_USER_ID.as("creator_user_id")),
                                     record.get(APP_USER.as("creator").LOGIN_ID.as("creator_login_id")),
                                     creatorRole));
 
@@ -91,11 +91,11 @@ public class JooqModuleSetReleaseReadRepository
             moduleSetRelease.setLastUpdatedBy(
                     (isUpdaterAdmin) ?
                             new ScoreUser(
-                                    record.get(APP_USER.as("updater").APP_USER_ID.as("updater_user_id")).toBigInteger(),
+                                    record.get(APP_USER.as("updater").APP_USER_ID.as("updater_user_id")),
                                     record.get(APP_USER.as("updater").LOGIN_ID.as("updater_login_id")),
                                     Arrays.asList(updaterRole, ADMINISTRATOR)) :
                             new ScoreUser(
-                                    record.get(APP_USER.as("updater").APP_USER_ID.as("updater_user_id")).toBigInteger(),
+                                    record.get(APP_USER.as("updater").APP_USER_ID.as("updater_user_id")),
                                     record.get(APP_USER.as("updater").LOGIN_ID.as("updater_login_id")),
                                     updaterRole));
 
@@ -166,7 +166,7 @@ public class JooqModuleSetReleaseReadRepository
         }
 
         if (request.getReleaseId() != null) {
-            conditions.add(MODULE_SET_RELEASE.RELEASE_ID.eq(ULong.valueOf(request.getReleaseId())));
+            conditions.add(MODULE_SET_RELEASE.RELEASE_ID.eq(request.getReleaseId()));
         }
 
         if (request.getDefault() != null) {
@@ -230,7 +230,7 @@ public class JooqModuleSetReleaseReadRepository
                         and(MODULE_ACC_MANIFEST.ACC_MANIFEST_ID.eq(ACC_MANIFEST.ACC_MANIFEST_ID),
                             MODULE_ACC_MANIFEST.MODULE_SET_RELEASE_ID.eq(ULong.valueOf(request.getModuleSetReleaseId()))))
                 .where(and(
-                        ACC_MANIFEST.RELEASE_ID.eq(ULong.valueOf(request.getReleaseId())),
+                        ACC_MANIFEST.RELEASE_ID.eq(request.getReleaseId()),
                         ACC.OBJECT_CLASS_TERM.notEqual("Any Structured Content"),
                         MODULE_ACC_MANIFEST.MODULE_ACC_MANIFEST_ID.isNull()))
                 .fetchStream().map(e -> {
@@ -288,7 +288,7 @@ public class JooqModuleSetReleaseReadRepository
                         and(MODULE_ASCCP_MANIFEST.ASCCP_MANIFEST_ID.eq(ASCCP_MANIFEST.ASCCP_MANIFEST_ID),
                                 MODULE_ASCCP_MANIFEST.MODULE_SET_RELEASE_ID.eq(ULong.valueOf(request.getModuleSetReleaseId()))))
                 .where(and(
-                        ASCCP_MANIFEST.RELEASE_ID.eq(ULong.valueOf(request.getReleaseId())),
+                        ASCCP_MANIFEST.RELEASE_ID.eq(request.getReleaseId()),
                         ASCCP.PROPERTY_TERM.notEqual("Any Property"),
                         MODULE_ASCCP_MANIFEST.MODULE_ASCCP_MANIFEST_ID.isNull()))
                 .fetchStream().map(e -> {
@@ -337,7 +337,7 @@ public class JooqModuleSetReleaseReadRepository
                 .from(BCCP_MANIFEST)
                 .join(BCC_MANIFEST).on(BCCP_MANIFEST.BCCP_MANIFEST_ID.eq(BCC_MANIFEST.TO_BCCP_MANIFEST_ID))
                 .join(BCC).on(BCC_MANIFEST.BCC_ID.eq(BCC.BCC_ID))
-                .where(and(BCCP_MANIFEST.RELEASE_ID.eq(ULong.valueOf(request.getReleaseId())),
+                .where(and(BCCP_MANIFEST.RELEASE_ID.eq(request.getReleaseId()),
                         BCC.ENTITY_TYPE.eq(1)))
                 .fetchInto(ULong.class);
         return dslContext().select(
@@ -353,7 +353,7 @@ public class JooqModuleSetReleaseReadRepository
                         and(MODULE_BCCP_MANIFEST.BCCP_MANIFEST_ID.eq(BCCP_MANIFEST.BCCP_MANIFEST_ID),
                                 MODULE_BCCP_MANIFEST.MODULE_SET_RELEASE_ID.eq(ULong.valueOf(request.getModuleSetReleaseId()))))
                 .where(and(
-                        BCCP_MANIFEST.RELEASE_ID.eq(ULong.valueOf(request.getReleaseId())),
+                        BCCP_MANIFEST.RELEASE_ID.eq(request.getReleaseId()),
                         BCCP_MANIFEST.BCCP_MANIFEST_ID.in(elementBccpManifestList),
                         MODULE_BCCP_MANIFEST.MODULE_BCCP_MANIFEST_ID.isNull()))
                 .fetchStream().map(e -> {
@@ -411,7 +411,7 @@ public class JooqModuleSetReleaseReadRepository
                         and(MODULE_DT_MANIFEST.DT_MANIFEST_ID.eq(DT_MANIFEST.DT_MANIFEST_ID),
                                 MODULE_DT_MANIFEST.MODULE_SET_RELEASE_ID.eq(ULong.valueOf(request.getModuleSetReleaseId()))))
                 .where(and(
-                        DT_MANIFEST.RELEASE_ID.eq(ULong.valueOf(request.getReleaseId())),
+                        DT_MANIFEST.RELEASE_ID.eq(request.getReleaseId()),
                         DT.BASED_DT_ID.isNotNull(),
                         MODULE_DT_MANIFEST.MODULE_DT_MANIFEST_ID.isNull()))
                 .fetchStream().map(e -> {
@@ -469,7 +469,7 @@ public class JooqModuleSetReleaseReadRepository
                         and(MODULE_CODE_LIST_MANIFEST.CODE_LIST_MANIFEST_ID.eq(CODE_LIST_MANIFEST.CODE_LIST_MANIFEST_ID),
                                 MODULE_CODE_LIST_MANIFEST.MODULE_SET_RELEASE_ID.eq(ULong.valueOf(request.getModuleSetReleaseId()))))
                 .where(and(
-                        CODE_LIST_MANIFEST.RELEASE_ID.eq(ULong.valueOf(request.getReleaseId())),
+                        CODE_LIST_MANIFEST.RELEASE_ID.eq(request.getReleaseId()),
                         MODULE_CODE_LIST_MANIFEST.MODULE_CODE_LIST_MANIFEST_ID.isNull()))
                 .fetchStream().map(e -> {
                     AssignableNode node = new AssignableNode();
@@ -526,7 +526,7 @@ public class JooqModuleSetReleaseReadRepository
                         and(MODULE_AGENCY_ID_LIST_MANIFEST.AGENCY_ID_LIST_MANIFEST_ID.eq(AGENCY_ID_LIST_MANIFEST.AGENCY_ID_LIST_MANIFEST_ID),
                                 MODULE_AGENCY_ID_LIST_MANIFEST.MODULE_SET_RELEASE_ID.eq(ULong.valueOf(request.getModuleSetReleaseId()))))
                 .where(and(
-                        AGENCY_ID_LIST_MANIFEST.RELEASE_ID.eq(ULong.valueOf(request.getReleaseId())),
+                        AGENCY_ID_LIST_MANIFEST.RELEASE_ID.eq(request.getReleaseId()),
                         MODULE_AGENCY_ID_LIST_MANIFEST.MODULE_AGENCY_ID_LIST_MANIFEST_ID.isNull()))
                 .fetchStream().map(e -> {
                     AssignableNode node = new AssignableNode();
@@ -582,7 +582,7 @@ public class JooqModuleSetReleaseReadRepository
                 .leftJoin(MODULE_XBT_MANIFEST).on(
                         and(MODULE_XBT_MANIFEST.XBT_MANIFEST_ID.eq(XBT_MANIFEST.XBT_MANIFEST_ID),
                                 MODULE_XBT_MANIFEST.MODULE_SET_RELEASE_ID.eq(ULong.valueOf(request.getModuleSetReleaseId()))))
-                .where(and(XBT_MANIFEST.RELEASE_ID.eq(ULong.valueOf(request.getReleaseId())),
+                .where(and(XBT_MANIFEST.RELEASE_ID.eq(request.getReleaseId()),
                         MODULE_XBT_MANIFEST.MODULE_XBT_MANIFEST_ID.isNull(),
                         XBT.BUILTIN_TYPE.notLike("xsd:%")))
                 .fetchStream().map(e -> {
